Added all policies

+++ app/controllers/posts_controller.rb
@@ -10,15 +10,18 @@ class PostsController < ApplicationController
   def show
     @comment = Comment.new
     @comments = Comment.where(post_id: @post.id)
+    authorize @comment
   end
 
   def new
     @post = Post.new
+    authorize @post
   end
 
   def create
     @post = Post.new(post_params)
     @post.user = current_user
+    authorize @post
     if @post.save
       redirect_to post_path(@post)
     else
@@ -27,14 +30,17 @@ class PostsController < ApplicationController
   end
 
   def edit
+    authorize @post
   end
 
   def update
+    authorize @post
     @post.update(post_params)
     redirect_to post_path(@post)
   end
 
   def destroy
+    authorize @post
     @post.destroy
     redirect_to user_path(current_user), data: { turbo_method: :delete, turbo_confirm: "Are you sure?" }
   end
